first iteration of input routine fixes

+++ SOURCE/FDISK/KBDINPUT.C
@@ -70,9 +70,9 @@ $set 4
 
 /* Get input from keyboard */
 unsigned long Input( int size_of_field, int x_position, int y_position,
-                     int type, int min_range, long max_range,
+                     int type, unsigned long min_range, unsigned long max_range,
                      int return_message, long default_value,
-                     long maximum_possible_percentage,
+                     unsigned long maximum_possible_percentage,
                      char optional_char_1[1], char optional_char_2[1] )
 {
    /*
@@ -637,7 +637,7 @@ unsigned long Input( int size_of_field, int x_position, int y_position,
                  ( input > 9 ) ) {
                proper_input_given = FALSE;
 
-               Color_Print_At( 4, 23, "Invalid entry, please enter %d-%d.",
+               Color_Print_At( 4, 23, "Invalid entry, please enter %lu-%lu.",
                                min_range, max_range );
                invalid_input = TRUE;
             }
@@ -649,7 +649,7 @@ unsigned long Input( int size_of_field, int x_position, int y_position,
 
                Color_Print_At( 4, 23, "%d is not a choice, please enter ",
                                input );
-               Color_Printf( "%d-%d.", min_range, max_range );
+               Color_Printf( "%lu-%lu.", min_range, max_range );
                invalid_input = TRUE;
             }
 
@@ -744,7 +744,7 @@ unsigned long Input( int size_of_field, int x_position, int y_position,
                    ( input > 9 ) ) ) {
                proper_input_given = FALSE;
 
-               Color_Print_At( 4, 23, "Invalid entry, please enter %d-%d.",
+               Color_Print_At( 4, 23, "Invalid entry, please enter %lu-%lu.",
                                min_range, max_range );
                invalid_input = TRUE;
             }
@@ -756,7 +756,7 @@ unsigned long Input( int size_of_field, int x_position, int y_position,
 
                Color_Print_At( 4, 23, "%d is not a choice, please enter ",
                                input );
-               Color_Printf( "%d-%d.", min_range, max_range );
+               Color_Printf( "%lu-%lu.", min_range, max_range );
                invalid_input = TRUE;
             }
